Move isNewer util out of LiveEvent
